ci: run invalid tests

+++ tests/gen_out_files.py
@@ -4,14 +4,14 @@
 
 # Script used to generate the `.out` files
 
-import os, glob, sys
-
-import utils
+import os, glob, sys, utils
 
 FAIL_FILES = glob.glob(f"tests/invalid/*.ri")
 for file in FAIL_FILES:
 	out_name = file.replace(".ri", ".out")
-	res = utils.run_process(sys.executable, "rivetc.py", file)
+	if os.path.exists(out_name):
+		continue
+	res = utils.run_process(sys.executable, "rivetc", file)
 	if res.exit_code != 0:
 		utils.eprint(f"[OK] {file}")
 		with open(out_name, "w", encoding = 'UTF-8') as f:
